Moved the auxtel_visit script functionality to a Class in a Python module.
Renamed ScriptQueue_Controller.py module to script_queue_controller.py.

Removed all the controller definitions from test_auxtel_visit.py. Now using the IntegrationTests ScriptQueueController Class. Testing that the integration test script is added to the queue.

Put the heartbeat loop into a try/except block.

Removed the remote.start_task call from aux_tel_visit.py module; it is redundant in the async with block.
