+++ bin/auxtel_visit.py
@@ -22,32 +22,8 @@
 
 import asyncio
 
-from lsst.ts import IntegrationTests
-from lsst.ts import salobj
-from lsst.ts.idl.enums import ScriptQueue
+from lsst.ts.IntegrationTests import AuxTelVisit
 
+script_class = AuxTelVisit()
 
-async def amain():
-    async with salobj.Domain() as domain, salobj.Remote(
-        domain=domain, name="ScriptQueue", index=2
-    ) as remote:
-        # note: use index=1 for MT, 2 for AuxTel;
-        # also since `async with` is used,
-        # you do NOT have to wait for the remote to start
-        configuration = IntegrationTests.auxtel_visit_config()
-        print("Test configuration:\n" + configuration)
-        await remote.start_task
-        await remote.evt_heartbeat.next(flush=True, timeout=30)
-        await remote.cmd_pause.start(timeout=10)
-        await remote.cmd_add.set_start(
-            timeout=10,
-            isStandard=True,
-            path="auxtel/take_image_latiss.py",
-            config=configuration,
-            logLevel=10,
-            location=ScriptQueue.Location.FIRST,
-        )
-        await remote.cmd_resume.set_start(timeout=10)
-
-
-asyncio.run(amain())
+asyncio.run(script_class.run())
