+++ python/lsst/ts/IntegrationTests/__init__.py
@@ -27,4 +27,5 @@ except ImportError:
 from .testutils import *
 from .yaml_test_strings import *
 from .configs import *
-from .ScriptQueue_Controller import *
+from .aux_tel_visit import *
+from .script_queue_controller import *
