+++ tests/test_black.py
@@ -1,4 +1,4 @@
-# This file is part of ts_salobj.
+# This file is part of ts_IntegrationTests.
 #
 # Developed for the Rubin Observatory Telescope and Site System.
 # This product includes software developed by the LSST Project
